+++ services/src/main/java/org/keycloak/broker/oidc/mappers/UserAttributeMapper.java
@@ -165,6 +165,10 @@ public class UserAttributeMapper extends AbstractClaimMapper {
             setIfNotEmpty(user::setFirstName, values);
         } else if (LAST_NAME.equalsIgnoreCase(attribute)) {
             setIfNotEmpty(user::setLastName, values);
+        } else if (EMAIL_VERIFIED.equalsIgnoreCase(attribute)) {
+            Object emailValue = getClaimValue(context, EMAIL);
+            if (values != null && ! values.isEmpty() && emailValue != null && user.getEmail().equals((String)emailValue ))
+                user.setEmailVerified(Boolean.valueOf(values.get(0)));
         } else {
             List<String> current = user.getAttributeStream(attribute).collect(Collectors.toList());
             if (!CollectionUtil.collectionEquals(values, current)) {